refactor(tui): reorganize source by TEA role into subdirectories
group sdk/tui modules by architectural role using modern Rust foo.rs + foo/
convention (view/, update/, model/, wizard/). public API and all
test-visible paths preserved via lib.rs re-exports; no test imports changed.

- view/: render modules (home, results, wizard, naming, shared, find)
- update/: command dispatcher and UpdateCtx (command.rs, ctx.rs)
- model/: state types (mode.rs, views.rs; formerly app_views.rs)
- wizard/: draft persistence (draft.rs; formerly wizard_draft.rs)
- wizard_common.rs: modernized from legacy wizard_common/mod.rs style

also fixes a license-header typo in view/naming.rs and updates
tests/budget.rs::no_async_in_render_path to recurse into src/view/.
