+++ sdk/tui/src/lib.rs
@@ -10,7 +10,6 @@
 
 pub mod app;
 pub mod app_launch;
-pub mod app_views;
 pub(crate) mod clipboard;
 pub mod command;
 pub mod find;
@@ -18,7 +17,6 @@ pub(crate) mod fuzzy;
 pub mod help;
 pub(crate) mod logo;
 pub mod message;
-pub mod mode;
 pub mod model;
 pub mod naming;
 pub mod runtime;
@@ -26,22 +24,19 @@ pub mod subscription;
 pub mod task;
 pub mod theme;
 pub mod update;
-pub(crate) mod update_command;
-pub mod update_ctx;
 pub mod view;
-pub(crate) mod view_find;
 pub mod wizard;
 pub mod wizard_common;
-pub mod wizard_draft;
+pub use wizard::draft as wizard_draft;
 
 pub use app_launch::{launch, LaunchOptions, ThemeChoice};
 pub use message::Message;
-pub use mode::{BrowsingState, ModalState, Mode, MouseMode, PaletteState};
+pub use model::mode::{BrowsingState, ModalState, Mode, MouseMode, PaletteState};
 pub use model::Model;
 pub use runtime::{dispatch, replay, run};
 pub use subscription::{subscriptions, Subscription, SubscriptionId, Subscriptions};
 pub use task::Task;
 pub use theme::Theme;
 pub use update::update;
-pub use update_ctx::{UpdateCtx, UpdateCtxBuilder};
+pub use update::ctx::{UpdateCtx, UpdateCtxBuilder};
 pub use view::draw;
