+++ sdk/tui/src/model/mode.rs
