+++ sdk/tui/src/view.rs
@@ -9,25 +9,29 @@
 // governing permissions and limitations under the License.
 
 use ratatui::{
-    layout::{Alignment, Constraint, Direction, Layout, Position, Rect},
-    style::{Modifier, Style},
+    layout::{Constraint, Direction, Layout, Rect},
+    style::Style,
     text::{Line, Span},
-    widgets::{Block, Borders, Cell, Clear, Paragraph, Row, Table},
+    widgets::{Block, Borders, Clear, Paragraph},
     Frame,
 };
 
-use crate::app::{
-    ActiveView, DescribeView, Modal, QueryView, ResolveView, StatusKind, ValidateView,
-};
-use crate::app_views::truncate_cell;
-use crate::command::Command;
+mod find;
+mod home;
+mod naming;
+mod results;
+pub(crate) mod shared;
+mod wizard;
+
+use crate::app::{ActiveView, Modal, StatusKind};
 use crate::help::HELP_TEXT;
-use crate::logo::LOGO;
 use crate::model::Model;
-use crate::naming::{NamingScreen, NamingWizardState};
 use crate::theme::Theme;
-use crate::view_find::render_find;
-use crate::wizard::{ClassificationDraft, ValueKind, WizardScreen, WizardState};
+use find::render_find;
+use home::render_home;
+use naming::render_naming;
+use results::{render_describe, render_query, render_resolve, render_validate};
+use wizard::render_wizard;
 
 /// Return a centered `Rect` covering `percent_x` × `percent_y` of `area`.
 fn centered_rect(percent_x: u16, percent_y: u16, area: Rect) -> Rect {
@@ -77,7 +81,7 @@ pub fn draw(model: &mut Model, frame: &mut Frame, theme: &Theme, primer_line: &s
     // Pre-extract palette fields before the active_view borrow.
     let palette_input: String = model.palette_input_value().to_string();
     let (palette_visual_cursor, palette_list_selected) =
-        if let crate::mode::Mode::InPalette(ref ps) = model.mode {
+        if let crate::model::mode::Mode::InPalette(ref ps) = model.mode {
             (ps.input.visual_cursor(), ps.list_selected)
         } else {
             (0, None)
@@ -141,239 +145,6 @@ pub fn draw(model: &mut Model, frame: &mut Frame, theme: &Theme, primer_line: &s
     }
 }
 
-// ── Per-view render fns (extracted from the inline match arms) ────────────────
-
-/// Render the home screen: always-on command palette, Obsidian-style.
-/// Layout: logo? · name+version · hint · separator · "> {input}" · filtered list.
-fn render_home(
-    frame: &mut Frame<'_>,
-    area: Rect,
-    theme: &Theme,
-    palette_input: &str,
-    palette_visual_cursor: usize,
-    list_selected: Option<usize>,
-) {
-    const MARGIN: &str = "  ";
-    const PROMPT_PREFIX: &str = "> ";
-
-    let filtered = Command::filter(palette_input);
-    // cmd_col: widest canonical name in the filtered set, for alignment.
-    let cmd_col = filtered.iter().map(|c| c.canonical().len()).max().unwrap_or(0);
-
-    let version = env!("CARGO_PKG_VERSION");
-    let logo_lines: Vec<&str> = LOGO.lines().collect();
-    // non_logo_height: name + hint + separator + prompt + up-to-8 commands + spacers
-    let non_logo_height: u16 = 13;
-    let show_logo = area.height >= logo_lines.len() as u16 + 1 + non_logo_height;
-
-    let mut lines: Vec<Line> = Vec::new();
-
-    if show_logo {
-        for l in &logo_lines {
-            lines.push(Line::from(format!("{MARGIN}{l}")));
-        }
-        lines.push(Line::from(""));
-    }
-
-    lines.push(Line::from(vec![
-        Span::raw(MARGIN),
-        Span::styled("Spectrum Design Data", Style::default().add_modifier(Modifier::BOLD)),
-        Span::styled(format!("  v{version}"), Style::default().fg(theme.muted)),
-    ]));
-    lines.push(Line::from(vec![
-        Span::raw(MARGIN),
-        Span::styled(
-            "↑↓ history/list · Tab complete · Enter run · Esc back · Ctrl+C quit",
-            Style::default().fg(theme.muted),
-        ),
-    ]));
-    lines.push(Line::from(vec![
-        Span::raw(MARGIN),
-        Span::styled(
-            "─".repeat(area.width.saturating_sub(MARGIN.len() as u16) as usize),
-            Style::default().fg(theme.muted),
-        ),
-    ]));
-    lines.push(Line::from(vec![
-        Span::raw(MARGIN),
-        Span::raw(PROMPT_PREFIX),
-        Span::raw(palette_input),
-    ]));
-
-    for (i, cmd) in filtered.iter().enumerate() {
-        let name = cmd.canonical();
-        // Selection style: when list is focused (list_selected is Some), the
-        // highlighted row gets selection_bg. When input is focused, the top
-        // match is bolded as a subtle "this is what Enter will run" hint.
-        let (row_style, name_style) = if list_selected == Some(i) {
-            (
-                Style::default().bg(theme.selection_bg),
-                Style::default()
-                    .fg(theme.accent)
-                    .bg(theme.selection_bg)
-                    .add_modifier(Modifier::BOLD),
-            )
-        } else if list_selected.is_none() && i == 0 {
-            (
-                Style::default(),
-                Style::default().fg(theme.accent).add_modifier(Modifier::BOLD),
-            )
-        } else {
-            (Style::default(), Style::default().fg(theme.accent))
-        };
-        let padding = " ".repeat(cmd_col.saturating_sub(name.len()) + 2);
-        let desc_style = if list_selected == Some(i) {
-            Style::default().fg(theme.muted).bg(theme.selection_bg)
-        } else {
-            Style::default().fg(theme.muted)
-        };
-        lines.push(
-            Line::from(vec![
-                Span::styled(format!("{MARGIN}  {name}"), name_style),
-                Span::styled(padding, row_style),
-                Span::styled(cmd.description(), desc_style),
-            ])
-            .style(row_style),
-        );
-    }
-
-    // Prompt row offset: logo+spacer (if shown) + name + hint + separator = N.
-    let prompt_offset: u16 = if show_logo { logo_lines.len() as u16 + 4 } else { 3 };
-    frame.render_widget(Paragraph::new(lines).alignment(Alignment::Left), area);
-    // Cursor always stays on the prompt line, even while browsing the list.
-    frame.set_cursor_position(Position {
-        x: area.x + MARGIN.len() as u16 + PROMPT_PREFIX.len() as u16 + palette_visual_cursor as u16,
-        y: area.y + prompt_offset,
-    });
-}
-
-fn render_query(f: &mut Frame<'_>, qv: &mut QueryView, area: Rect, theme: &Theme) {
-    let name_max = (area.width.saturating_sub(5) as usize) * 40 / 100;
-    let header = Row::new(vec![
-        Cell::from("Name").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Value").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("File").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Layer").style(Style::default().add_modifier(Modifier::BOLD)),
-    ]);
-    let rows: Vec<Row> = qv
-        .rows
-        .iter()
-        .map(|r| {
-            Row::new(vec![
-                Cell::from(truncate_cell(&r.name, name_max)),
-                Cell::from(r.value.as_str()),
-                Cell::from(r.file.as_str()),
-                Cell::from(r.layer.as_str()),
-            ])
-        })
-        .collect();
-    let widths = [
-        Constraint::Percentage(40),
-        Constraint::Percentage(30),
-        Constraint::Percentage(20),
-        Constraint::Percentage(10),
-    ];
-    let title = if qv.is_fuzzy {
-        format!(" Fuzzy: /{} ", qv.expr_text)
-    } else {
-        format!(" Query: {} ", qv.expr_text)
-    };
-    let table = Table::new(rows, widths)
-        .header(header)
-        .block(Block::default().borders(Borders::ALL).title(title))
-        .row_highlight_style(Style::default().bg(theme.selection_bg));
-    f.render_stateful_widget(table, area, &mut qv.table_state);
-}
-
-fn render_resolve(f: &mut Frame<'_>, rv: &mut ResolveView, area: Rect, theme: &Theme) {
-    let name_max = (area.width.saturating_sub(9) as usize) * 35 / 100;
-    let header = Row::new(vec![
-        Cell::from("★").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Name").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Value").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("File").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Layer").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Spec").style(Style::default().add_modifier(Modifier::BOLD)),
-    ]);
-    let rows: Vec<Row> = rv
-        .rows
-        .iter()
-        .map(|r| {
-            Row::new(vec![
-                Cell::from(if r.is_winner { "★" } else { "" }),
-                Cell::from(truncate_cell(&r.name, name_max)),
-                Cell::from(r.value.as_str()),
-                Cell::from(r.file.as_str()),
-                Cell::from(r.layer.as_str()),
-                Cell::from(r.specificity.to_string()),
-            ])
-        })
-        .collect();
-    let widths = [
-        Constraint::Length(2),
-        Constraint::Percentage(35),
-        Constraint::Percentage(25),
-        Constraint::Percentage(20),
-        Constraint::Percentage(12),
-        Constraint::Percentage(8),
-    ];
-    let table = Table::new(rows, widths)
-        .header(header)
-        .block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title(format!(" Resolve: {} ", rv.property)),
-        )
-        .row_highlight_style(Style::default().bg(theme.selection_bg));
-    f.render_stateful_widget(table, area, &mut rv.table_state);
-}
-
-fn render_describe(f: &mut Frame<'_>, dv: &DescribeView, area: Rect) {
-    let para = Paragraph::new(dv.pretty_json.as_str())
-        .block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title(format!(" Describe: {} ", dv.component)),
-        )
-        .scroll((dv.scroll, 0));
-    f.render_widget(para, area);
-}
-
-fn render_validate(f: &mut Frame<'_>, vv: &mut ValidateView, area: Rect, theme: &Theme) {
-    let token_max = (area.width.saturating_sub(12) as usize) * 28 / 100;
-    let header = Row::new(vec![
-        Cell::from("Sev").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Rule").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Token").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Message").style(Style::default().add_modifier(Modifier::BOLD)),
-    ]);
-    let rows: Vec<Row> = vv
-        .rows
-        .iter()
-        .map(|r| {
-            Row::new(vec![
-                Cell::from(r.severity.as_str()),
-                Cell::from(r.rule_id.as_str()),
-                Cell::from(truncate_cell(&r.token, token_max)),
-                Cell::from(r.message.as_str()),
-            ])
-        })
-        .collect();
-    let widths = [
-        Constraint::Length(7),
-        Constraint::Percentage(12),
-        Constraint::Percentage(28),
-        Constraint::Percentage(60),
-    ];
-    let table = Table::new(rows, widths)
-        .header(header)
-        .block(Block::default().borders(Borders::ALL).title(" Validate "))
-        .row_highlight_style(Style::default().bg(theme.selection_bg));
-    f.render_stateful_widget(table, area, &mut vv.table_state);
-}
-
-// ── Help modal ────────────────────────────────────────────────────────────────
-
 fn render_help_modal(f: &mut Frame<'_>, scroll: u16, area: Rect) {
     let popup_area = centered_rect(80, 90, area);
     f.render_widget(Clear, popup_area);
@@ -385,416 +156,4 @@ fn render_help_modal(f: &mut Frame<'_>, scroll: u16, area: Rect) {
         )
         .scroll((scroll, 0));
     f.render_widget(para, popup_area);
-}
-
-// ── Wizard render ─────────────────────────────────────────────────────────────
-
-fn render_wizard(f: &mut Frame<'_>, ws: &mut WizardState, area: Rect, theme: &Theme) {
-    let screen_num = ws.screen.number();
-    let screen_name = ws.screen.name();
-
-    let outer = Block::default()
-        .borders(Borders::ALL)
-        .title(format!(" Wizard · {screen_num}/4 · {screen_name} "));
-    let inner_area = outer.inner(area);
-    f.render_widget(outer, area);
-
-    let inner_chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Min(0), Constraint::Length(1)])
-        .split(inner_area);
-
-    let footer_text = match ws.screen {
-        WizardScreen::Intent => {
-            "Enter: continue  Tab: reuse selected  ↑↓: select suggestion  Esc: cancel"
-        }
-        WizardScreen::Classification => {
-            "Tab/Shift-Tab: next/prev field  ←→: cycle layer  +: add name field  Enter: continue  Esc: cancel"
-        }
-        WizardScreen::Values => {
-            "a: alias  l: literal  e: edit value  ↑↓: select row  Enter: continue  Esc: cancel"
-        }
-        WizardScreen::Confirm => {
-            "Type rationale, then Enter to submit  ↑↓: scroll diff  Ctrl+S: edit $schema  Esc: cancel"
-        }
-    };
-    f.render_widget(
-        Paragraph::new(footer_text).style(Style::default().fg(theme.muted)),
-        inner_chunks[1],
-    );
-
-    match ws.screen {
-        WizardScreen::Intent => render_intent_screen(f, ws, inner_chunks[0], theme),
-        WizardScreen::Classification => render_classification_screen(f, ws, inner_chunks[0]),
-        WizardScreen::Values => render_values_screen(f, ws, inner_chunks[0], theme),
-        WizardScreen::Confirm => render_confirm_screen(f, ws, inner_chunks[0], theme),
-    }
-}
-
-fn render_intent_screen(f: &mut Frame<'_>, ws: &WizardState, area: Rect, theme: &Theme) {
-    render_intent_content(
-        f,
-        ws.can_alias,
-        ws.intent.value(),
-        &ws.suggestions,
-        ws.selected_suggestion,
-        area,
-        theme,
-    );
-}
-
-fn render_classification_screen(f: &mut Frame<'_>, ws: &WizardState, area: Rect) {
-    render_classification_content(f, &ws.classification, &ws.assembled_name(), area);
-}
-
-// ── Shared widget render helpers ──────────────────────────────────────────────
-
-fn render_intent_content(
-    f: &mut Frame<'_>,
-    can_alias: bool,
-    intent_value: &str,
-    suggestions: &[design_data_core::suggest::SuggestionResult],
-    selected_suggestion: usize,
-    area: Rect,
-    theme: &Theme,
-) {
-    let banner_height: u16 = if can_alias { 3 } else { 0 };
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(1),
-            Constraint::Length(banner_height),
-            Constraint::Min(0),
-        ])
-        .split(area);
-
-    let intent_line = format!("Intent: {intent_value}");
-    f.render_widget(Paragraph::new(intent_line), chunks[0]);
-
-    if can_alias {
-        let accent = Style::default()
-            .fg(theme.accent)
-            .add_modifier(Modifier::BOLD);
-        let banner = Paragraph::new(vec![
-            Line::from(Span::styled(
-                "  These tokens already exist for similar intents.",
-                accent,
-            )),
-            Line::from(Span::styled(
-                "  Reusing one keeps the cascade healthy.  Tab to alias · Enter to create new",
-                Style::default().fg(theme.accent),
-            )),
-        ]);
-        f.render_widget(banner, chunks[1]);
-    }
-
-    let list_area = chunks[2];
-    if suggestions.is_empty() {
-        if !intent_value.is_empty() {
-            f.render_widget(
-                Paragraph::new("  (no suggestions — will create new token)"),
-                list_area,
-            );
-        } else {
-            f.render_widget(
-                Paragraph::new("  Type to search for existing tokens…"),
-                list_area,
-            );
-        }
-    } else {
-        let sources: Vec<String> = suggestions
-            .iter()
-            .map(|s| {
-                s.file
-                    .file_name()
-                    .and_then(|n| n.to_str())
-                    .unwrap_or("")
-                    .trim_end_matches(".tokens.json")
-                    .to_string()
-            })
-            .collect();
-        let source_col_width = sources
-            .iter()
-            .map(|s| s.len())
-            .max()
-            .unwrap_or(0)
-            .min(24) as u16;
-        let rows: Vec<Row> = suggestions
-            .iter()
-            .enumerate()
-            .zip(sources.iter())
-            .map(|((i, s), source)| {
-                let marker = if i == selected_suggestion { "▶" } else { " " };
-                let conf = format!("{:.0}%", s.confidence * 100.0);
-                Row::new(vec![
-                    Cell::from(marker),
-                    Cell::from(s.display_name()),
-                    Cell::from(Span::styled(
-                        source.clone(),
-                        Style::default().fg(theme.muted),
-                    )),
-                    Cell::from(conf),
-                ])
-            })
-            .collect();
-        let widths = [
-            Constraint::Length(2),
-            Constraint::Min(0),
-            Constraint::Length(source_col_width),
-            Constraint::Length(5),
-        ];
-        let table =
-            Table::new(rows, widths).row_highlight_style(Style::default().bg(theme.selection_bg));
-        f.render_widget(table, list_area);
-    }
-}
-
-fn render_classification_content(
-    f: &mut Frame<'_>,
-    classification: &ClassificationDraft,
-    assembled_name: &str,
-    area: Rect,
-) {
-    let layer_str = match classification.layer {
-        design_data_core::graph::Layer::Foundation => "Foundation",
-        design_data_core::graph::Layer::Platform => "Platform",
-        design_data_core::graph::Layer::Product => "Product",
-    };
-
-    let mut lines: Vec<Line> = Vec::new();
-    let focused = classification.focused_field;
-
-    let layer_label = if focused == 0 {
-        format!("▶ Layer:    ← {layer_str} →")
-    } else {
-        format!("  Layer:      {layer_str}")
-    };
-    lines.push(Line::from(layer_label));
-
-    let prop_label = if focused == 1 {
-        format!("▶ Property: {}", classification.property.value())
-    } else {
-        format!("  Property: {}", classification.property.value())
-    };
-    lines.push(Line::from(prop_label));
-
-    for (i, field) in classification.name_fields.iter().enumerate() {
-        let marker = if focused == i + 2 { "▶" } else { " " };
-        lines.push(Line::from(format!(
-            "{marker} {}: {}",
-            field.key,
-            field.value.value()
-        )));
-    }
-
-    lines.push(Line::from(""));
-    lines.push(Line::from(format!("  Preview: {assembled_name}")));
-
-    f.render_widget(Paragraph::new(lines), area);
-}
-
-// ── Naming wizard render ──────────────────────────────────────────────────────
-
-fn render_naming(f: &mut Frame<'_>, ns: &NamingWizardState, area: Rect, theme: &Theme) {
-    let screen_num = ns.screen.number();
-    let screen_name = ns.screen.name();
-
-    let outer = Block::default().borders(Borders::ALL).title(format!(
-        " Name · {screen_num}/{} · {screen_name} ",
-        NamingScreen::SCREEN_COUNT
-    ));
-    let inner_area = outer.inner(area);
-    f.render_widget(outer, area);
-
-    let inner_chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Min(0), Constraint::Length(1)])
-        .split(inner_area);
-
-    let footer_text = match ns.screen {
-        NamingScreen::Intent => "Enter: continue  ↑↓: select suggestion  Esc: cancel",
-        NamingScreen::Classification => {
-            "Tab/Shift-Tab: next/prev field  ←→/b: cycle layer / back to intent  +: add name field  Enter: done  Esc: cancel"
-        }
-        NamingScreen::Result => "c/y: copy name  e: edit  Esc/q: close",
-    };
-    f.render_widget(
-        Paragraph::new(footer_text).style(Style::default().fg(theme.muted)),
-        inner_chunks[1],
-    );
-
-    match ns.screen {
-        NamingScreen::Intent => render_intent_content(
-            f,
-            ns.can_alias,
-            ns.intent.value(),
-            &ns.suggestions,
-            ns.selected_suggestion,
-            inner_chunks[0],
-            theme,
-        ),
-        NamingScreen::Classification => render_classification_content(
-            f,
-            &ns.classification,
-            &ns.assembled_name(),
-            inner_chunks[0],
-        ),
-        NamingScreen::Result => render_naming_result(f, ns, inner_chunks[0], theme),
-    }
-}
-
-fn render_naming_result(f: &mut Frame<'_>, ns: &NamingWizardState, area: Rect, theme: &Theme) {
-    let name = ns.assembled_name();
-    let display = if name.is_empty() {
-        "(no name assembled — go back and fill in Property)".to_string()
-    } else {
-        name
-    };
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Length(3), Constraint::Min(0)])
-        .split(area);
-
-    let name_block = Block::default()
-        .borders(Borders::ALL)
-        .title(" Assembled name ");
-    let name_para = Paragraph::new(Span::styled(
-        display,
-        Style::default()
-            .fg(theme.accent)
-            .add_modifier(Modifier::BOLD),
-    ))
-    .block(name_block);
-    f.render_widget(name_para, chunks[0]);
-
-    let hint = Paragraph::new(vec![
-        Line::from("  Press c or y to copy this name to the clipboard."),
-        Line::from("  Press e to go back and refine the classification."),
-        Line::from("  Press Esc or q to close without copying."),
-    ])
-    .style(Style::default().fg(theme.muted));
-    f.render_widget(hint, chunks[1]);
-}
-
-// ── Wizard screen renders ─────────────────────────────────────────────────────
-
-fn render_values_screen(f: &mut Frame<'_>, ws: &WizardState, area: Rect, theme: &Theme) {
-    if ws.values.rows.is_empty() {
-        f.render_widget(
-            Paragraph::new("  (no mode combinations — graph has no mode sets)"),
-            area,
-        );
-        return;
-    }
-
-    let header = Row::new(vec![
-        Cell::from("Mode combo").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Kind").style(Style::default().add_modifier(Modifier::BOLD)),
-        Cell::from("Value / Alias target").style(Style::default().add_modifier(Modifier::BOLD)),
-    ]);
-    let rows: Vec<Row> = ws
-        .values
-        .rows
-        .iter()
-        .enumerate()
-        .map(|(i, row)| {
-            let combo = row.combo_label();
-            let kind = match row.kind {
-                ValueKind::Alias => "alias",
-                ValueKind::Literal => "literal",
-            };
-            let value = match row.kind {
-                ValueKind::Alias => row.alias_target.value().to_string(),
-                ValueKind::Literal => row.literal.value().to_string(),
-            };
-            let style = if i == ws.values.selected {
-                Style::default().bg(theme.selection_bg)
-            } else {
-                Style::default()
-            };
-            Row::new(vec![Cell::from(combo), Cell::from(kind), Cell::from(value)]).style(style)
-        })
-        .collect();
-
-    let widths = [
-        Constraint::Percentage(30),
-        Constraint::Percentage(10),
-        Constraint::Percentage(60),
-    ];
-    let table = Table::new(rows, widths)
-        .header(header)
-        .block(Block::default().borders(Borders::NONE));
-    f.render_widget(table, area);
-}
-
-fn render_confirm_screen(f: &mut Frame<'_>, ws: &WizardState, area: Rect, theme: &Theme) {
-    let rationale_height = 3u16;
-    let error_height = if ws.error.is_some() { 1u16 } else { 0u16 };
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(1), // schema URL / editor
-            Constraint::Length(rationale_height),
-            Constraint::Min(0),               // diff preview
-            Constraint::Length(error_height), // write error
-        ])
-        .split(area);
-
-    // Schema URL header / inline editor.
-    let schema_line = if ws.editing_schema_url {
-        Line::from(vec![
-            Span::styled("  $schema: ", Style::default().fg(theme.accent)),
-            Span::raw(ws.schema_url_input.value()),
-            Span::styled("▌", Style::default().fg(theme.accent)),
-        ])
-    } else {
-        let url_text = ws.schema_url.as_deref().unwrap_or("(none — Ctrl+S to set)");
-        Line::from(vec![
-            Span::styled("  $schema: ", Style::default().fg(theme.muted)),
-            Span::raw(url_text),
-        ])
-    };
-    f.render_widget(Paragraph::new(schema_line), chunks[0]);
-
-    // Rationale input.
-    let rationale_block = Block::default()
-        .borders(Borders::ALL)
-        .title(" Rationale (required) ");
-    let rationale_text = ws.rationale.value();
-    let rationale_line = if rationale_text.len() > 280 {
-        Line::from(vec![
-            Span::raw(rationale_text),
-            Span::styled(" ⚠ >280 chars", Style::default().fg(theme.warn)),
-        ])
-    } else {
-        Line::from(Span::raw(rationale_text))
-    };
-    f.render_widget(
-        Paragraph::new(rationale_line).block(rationale_block),
-        chunks[1],
-    );
-
-    // Diff preview.
-    let diff_text = ws
-        .diff_preview
-        .as_deref()
-        .unwrap_or("(diff will appear here once rationale is added and Enter pressed)");
-    let diff_para = Paragraph::new(diff_text)
-        .block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title(" Diff preview "),
-        )
-        .scroll((ws.diff_scroll, 0));
-    f.render_widget(diff_para, chunks[2]);
-
-    // Write error.
-    if let Some(ref err) = ws.error {
-        f.render_widget(
-            Paragraph::new(format!("  ⚠ {err}")).style(Style::default().fg(theme.error)),
-            chunks[3],
-        );
-    }
 }
