+++ sdk/tui/src/view/find.rs
